refactor: сÑchange dispatcher call location
