+++ src/components/App/App.jsx
@@ -1,23 +1,12 @@
-import React, {useEffect} from 'react';
-import {useDispatch, useSelector} from 'react-redux';
+import React from 'react';
+import {useSelector} from 'react-redux';
 import styles from './App.module.css';
-import {addSelectedItem, getIngredients} from '../../store/ingredientsSlice';
 import AppHeader from '../AppHeader/AppHeader';
 import BurgerIngredients from '../BurgerIngredients/BurgerIngredients';
 import BurgerConstructor from '../BurgerConstructor/BurgerConstructor';
 
 function App() {
-  const dispatch = useDispatch();
   const ingredients = useSelector(state => state.ingredients.items);
-
-  useEffect(() => {
-    dispatch(addSelectedItem(ingredients[0]));
-  }, [ingredients])
-
-
-  useEffect(() => {
-    dispatch(getIngredients())
-  },[dispatch])
 
   return (
     <>
