+++ src/components/BurgerIngredients/BurgerIngredients.jsx
@@ -1,11 +1,24 @@
-import React, { useState } from 'react';
+import React, {useEffect, useState} from 'react';
 import styles from './BurgerIngredients.module.css';
 import Tabs from '../Tabs/Tabs';
 import IngredientList from '../IngredientList/IngredientList';
 import Modal from '../Modal/Modal';
 import IngredientDetails from '../IngredientDetails/IngredientDetails';
+import {useDispatch, useSelector} from 'react-redux';
+import {addSelectedItem, getIngredients} from '../../store/ingredientsSlice';
 
 export default function BurgerIngredients() {
+  const dispatch = useDispatch();
+  const ingredients = useSelector(state => state.ingredients.items);
+
+  useEffect(() => {
+    dispatch(getIngredients())
+  },[dispatch])
+
+  useEffect(() => {
+    dispatch(addSelectedItem(ingredients[0]));
+  }, [ingredients])
+
   const [selectedIngredient, setSelectedIngredient] = useState({
     isPopupVisible: false,
     value: false
